Clear the warning of battery_charging_userstate_t; Add function Paytm_MQTT_Disconnect_Only which just disconnect from server but not deinit; Add annotations for the timer function

+++ Code/sdk/app/paytm_api/inc/paytm_dev_api.h
@@ -23,7 +23,7 @@ int64 Paytm_GetTotalHeapSize(void);
  * @description: Return free ram size of device, uints is byte
  * @return {*}
  */
-int64 Paytm_GetFreeHeapSize(void);
+int32 Paytm_GetFreeHeapSize(void);
 int64 Paytm_GetPhysicalRAM(void);
 int64 Paytm_GetTotalRAM(void);
